Everything is working, yuppie!

+++ src/main/java/common/messages/FileUpdateMessage.java
@@ -2,4 +2,4 @@ package common.messages;
 
 import java.io.Serializable;
 
-public record FileUpdateMessage(String userID, String username, String filename, long date, double size) implements Serializable {}
+public record FileUpdateMessage(String userID, String username, String filename, long date, long size) implements Serializable {}
